+++ crates/nix_rs/src/command.rs
@@ -8,7 +8,10 @@
 //! cmd.run_with_args_returning_stdout(&["--version"]);
 //! ```
 
-use std::fmt::{self, Display};
+use std::{
+    ffi::OsStr,
+    fmt::{self, Display},
+};
 
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
@@ -148,7 +151,11 @@ impl NixCmd {
     }
 
     /// Run nix with given args, letting stdout and stderr be that of parent process
-    pub async fn run_with_args(&self, args: &[&str]) -> Result<(), CommandError> {
+    pub async fn run_with_args<I, S>(&self, args: I) -> Result<(), CommandError>
+    where
+        I: IntoIterator<Item = S>,
+        S: AsRef<OsStr>,
+    {
         let mut cmd = self.command();
         cmd.args(args);
         trace_cmd(&cmd);